+++ contracts/ClearingHouse.sol
@@ -66,12 +66,14 @@ contract ClearingHouse is
         bool isClose;
         uint256 amount;
         uint160 sqrtPriceLimitX96;
+        bool isLiquidation;
     }
 
     struct InternalClosePositionParams {
         address trader;
         address baseToken;
         uint160 sqrtPriceLimitX96;
+        bool isLiquidation;
     }
 
     struct InternalCheckSlippageParams {
@@ -417,7 +419,8 @@ contract ClearingHouse is
                     isExactInput: params.isExactInput,
                     amount: params.amount,
                     isClose: false,
-                    sqrtPriceLimitX96: params.sqrtPriceLimitX96
+                    sqrtPriceLimitX96: params.sqrtPriceLimitX96,
+                    isLiquidation: false
                 })
             );
 
@@ -465,16 +468,14 @@ contract ClearingHouse is
                 InternalClosePositionParams({
                     trader: trader,
                     baseToken: params.baseToken,
-                    sqrtPriceLimitX96: params.sqrtPriceLimitX96
+                    sqrtPriceLimitX96: params.sqrtPriceLimitX96,
+                    isLiquidation: false
                 })
             );
 
         // if exchangedPositionSize < 0, closing it is short, B2Q; else, closing it is long, Q2B
         bool isBaseToQuote = response.exchangedPositionSize < 0 ? true : false;
-        uint256 oppositeAmountBound =
-            response.isPartialClose
-                ? params.oppositeAmountBound.mulRatio(IClearingHouseConfig(_clearingHouseConfig).getPartialCloseRatio())
-                : params.oppositeAmountBound;
+        uint256 oppositeAmountBound = _getPartialOppositeAmount(params.oppositeAmountBound, response.isPartialClose);
 
         _checkSlippage(
             InternalCheckSlippageParams({
@@ -493,7 +494,51 @@ contract ClearingHouse is
     }
 
     /// @inheritdoc IClearingHouse
-    function liquidate(address trader, address baseToken) external override whenNotPaused nonReentrant {
+    function liquidate(
+        address trader,
+        address baseToken,
+        uint256 oppositeAmountBound
+    )
+        external
+        override
+        returns (
+            uint256 base,
+            uint256 quote,
+            bool isPartialClose
+        )
+    {
+        int256 positionSize = IAccountBalance(_accountBalance).getTakerPositionSize(trader, baseToken);
+        // if positionSize > 0, it's long, and closing it is thus short, B2Q; else, closing it is long, Q2B
+        bool isBaseToQuote = positionSize > 0;
+
+        (base, quote, isPartialClose) = liquidate(trader, baseToken);
+
+        oppositeAmountBound = _getPartialOppositeAmount(oppositeAmountBound, isPartialClose);
+        _checkSlippage(
+            InternalCheckSlippageParams({
+                isBaseToQuote: isBaseToQuote,
+                isExactInput: isBaseToQuote,
+                base: base,
+                quote: quote,
+                oppositeAmountBound: oppositeAmountBound
+            })
+        );
+
+        return (base, quote, isPartialClose);
+    }
+
+    /// @inheritdoc IClearingHouse
+    function liquidate(address trader, address baseToken)
+        public
+        override
+        whenNotPaused
+        nonReentrant
+        returns (
+            uint256 base,
+            uint256 quote,
+            bool isPartialClose
+        )
+    {
         // liquidation trigger:
         //   accountMarginRatio < accountMaintenanceMarginRatio
         //   => accountValue / sum(abs(positionValue_market)) <
@@ -520,7 +565,14 @@ contract ClearingHouse is
         // must settle funding first
         _settleFunding(trader, baseToken);
         IExchange.SwapResponse memory response =
-            _closePosition(InternalClosePositionParams({ trader: trader, baseToken: baseToken, sqrtPriceLimitX96: 0 }));
+            _closePosition(
+                InternalClosePositionParams({
+                    trader: trader,
+                    baseToken: baseToken,
+                    sqrtPriceLimitX96: 0,
+                    isLiquidation: true
+                })
+            );
 
         // trader's pnl-- as liquidation penalty
         uint256 liquidationFee =
@@ -542,6 +594,8 @@ contract ClearingHouse is
             liquidationFee,
             liquidator
         );
+
+        return (response.base, response.quote, response.isPartialClose);
     }
 
     /// @inheritdoc IClearingHouse
@@ -839,6 +893,17 @@ contract ClearingHouse is
                 params.baseToken,
                 response.pnlToBeRealized
             );
+
+            // if realized pnl is not zero, that means trader is reducing or closing position
+            // trader cannot reduce/close position if bad debt happen
+            // unless it's a liquidation from backstop liquidity provider
+            // CH_BD: trader has bad debt after reducing/closing position
+            require(
+                (params.isLiquidation &&
+                    IClearingHouseConfig(_clearingHouseConfig).isBackstopLiquidityProvider(_msgSender())) ||
+                    getAccountValue(params.trader) >= 0,
+                "CH_BD"
+            );
         }
 
         // if not closing a position, check margin ratio after swap
@@ -873,19 +938,19 @@ contract ClearingHouse is
         // CH_PSZ: position size is zero
         require(positionSize != 0, "CH_PSZ");
 
-        // old position is long. when closing, it's baseToQuote && exactInput (sell exact base)
-        // old position is short. when closing, it's quoteToBase && exactOutput (buy exact base back)
-        bool isLong = positionSize > 0;
+        // if positionSize > 0, it's long, and closing it is thus short, B2Q; else, closing it is long, Q2B
+        bool isBaseToQuote = positionSize > 0;
         return
             _openPosition(
                 InternalOpenPositionParams({
                     trader: params.trader,
                     baseToken: params.baseToken,
-                    isBaseToQuote: isLong,
-                    isExactInput: isLong,
+                    isBaseToQuote: isBaseToQuote,
+                    isExactInput: isBaseToQuote,
                     isClose: true,
                     amount: positionSize.abs(),
-                    sqrtPriceLimitX96: params.sqrtPriceLimitX96
+                    sqrtPriceLimitX96: params.sqrtPriceLimitX96,
+                    isLiquidation: params.isLiquidation
                 })
             );
     }
@@ -935,6 +1000,17 @@ contract ClearingHouse is
             _getFreeCollateralByRatio(trader, IClearingHouseConfig(_clearingHouseConfig).getImRatio()) >= 0,
             "CH_NEFCI"
         );
+    }
+
+    function _getPartialOppositeAmount(uint256 oppositeAmountBound, bool isPartialClose)
+        internal
+        view
+        returns (uint256)
+    {
+        return
+            isPartialClose
+                ? oppositeAmountBound.mulRatio(IClearingHouseConfig(_clearingHouseConfig).getPartialCloseRatio())
+                : oppositeAmountBound;
     }
 
     function _checkSlippage(InternalCheckSlippageParams memory params) internal pure {