+++ test/clearingHouse/ClearingHouse.closePosition.test.ts
@@ -13,7 +13,8 @@ import {
     Vault,
 } from "../../typechain"
 import { addOrder, closePosition, q2bExactInput, q2bExactOutput } from "../helper/clearingHouseHelper"
-import { getMaxTick, getMinTick } from "../helper/number"
+import { initAndAddPool } from "../helper/marketHelper"
+import { getMaxTick, getMaxTickRange, getMinTick } from "../helper/number"
 import { deposit } from "../helper/token"
 import { forwardTimestamp } from "../shared/time"
 import { encodePriceSqrt } from "../shared/utilities"
@@ -70,12 +71,15 @@ describe("ClearingHouse closePosition", () => {
 
     describe("one maker; initialized price = 151.373306858723226652", () => {
         beforeEach(async () => {
-            await pool.initialize(encodePriceSqrt("151.373306858723226652", "1")) // tick = 50200 (1.0001^50200 = 151.373306858723226652)
-            // the initial number of oracle can be recorded is 1; thus, have to expand it
-            await pool.increaseObservationCardinalityNext((2 ^ 16) - 1)
-
-            // add pool after it's initialized
-            await marketRegistry.addPool(baseToken.address, 10000)
+            await initAndAddPool(
+                fixture,
+                pool,
+                baseToken.address,
+                encodePriceSqrt("151.373306858723226652", "1"), // tick = 50200 (1.0001^50200 = 151.373306858723226652)
+                10000,
+                // set maxTickCrossed as maximum tick range of pool by default, that means there is no over price when swap
+                getMaxTickRange(),
+            )
 
             // alice add liquidity
             const addLiquidityParams = {
@@ -253,12 +257,15 @@ describe("ClearingHouse closePosition", () => {
     // different range
     describe("two makers; initialized price = 148.3760629", () => {
         beforeEach(async () => {
-            await pool.initialize(encodePriceSqrt(148.3760629, 1))
-            // the initial number of oracle can be recorded is 1; thus, have to expand it
-            await pool.increaseObservationCardinalityNext((2 ^ 16) - 1)
-
-            // add pool after it's initialized
-            await marketRegistry.addPool(baseToken.address, 10000)
+            await initAndAddPool(
+                fixture,
+                pool,
+                baseToken.address,
+                encodePriceSqrt(148.3760629, 1),
+                10000,
+                // set maxTickCrossed as maximum tick range of pool by default, that means there is no over price when swap
+                getMaxTickRange(),
+            )
         })
 
         it("ranges of makers are the same; alice receives 3/4 of fee, while carol receives only 1/4", async () => {
@@ -630,12 +637,15 @@ describe("ClearingHouse closePosition", () => {
     describe("dust test for UniswapV3Broker.swap()", () => {
         // this test will fail if the _DUST constant in UniswapV3Broker is set to 1 (no dust allowed)
         it("a trader swaps base to quote and then closes; one maker", async () => {
-            await pool.initialize(encodePriceSqrt(151.3733069, 1))
-            // the initial number of oracle can be recorded is 1; thus, have to expand it
-            await pool.increaseObservationCardinalityNext((2 ^ 16) - 1)
-
-            // add pool after it's initialized
-            await marketRegistry.addPool(baseToken.address, 10000)
+            await initAndAddPool(
+                fixture,
+                pool,
+                baseToken.address,
+                encodePriceSqrt(151.3733069, 1),
+                10000,
+                // set maxTickCrossed as maximum tick range of pool by default, that means there is no over price when swap
+                getMaxTickRange(),
+            )
 
             // alice add liquidity
             const addLiquidityParams = {
@@ -683,14 +693,16 @@ describe("ClearingHouse closePosition", () => {
     describe("close position when user is maker and taker", async () => {
         let lowerTick: number, upperTick: number
         beforeEach(async () => {
-            await pool.initialize(encodePriceSqrt(151.3733069, 1))
-            // the initial number of oracle can be recorded is 1; thus, have to expand it
-            await pool.increaseObservationCardinalityNext((2 ^ 16) - 1)
-
-            // add pool after it's initialized
-            await marketRegistry.addPool(baseToken.address, 10000)
-
             const tickSpacing = await pool.tickSpacing()
+            await initAndAddPool(
+                fixture,
+                pool,
+                baseToken.address,
+                encodePriceSqrt(151.3733069, 1),
+                10000,
+                // set maxTickCrossed as maximum tick range of pool by default, that means there is no over price when swap
+                getMaxTickRange(),
+            )
             lowerTick = getMinTick(tickSpacing)
             upperTick = getMaxTick(tickSpacing)
         })