+++ test/clearingHouse/fixtures.ts
@@ -151,8 +151,8 @@ export function createClearingHouseFixture(
         let clearingHouse: ClearingHouse | TestClearingHouse
         if (canMockTime) {
             const clearingHouseFactory = await ethers.getContractFactory("TestClearingHouse")
-            clearingHouse = (await clearingHouseFactory.deploy()) as TestClearingHouse
-            await clearingHouse.initialize(
+            const testClearingHouse = (await clearingHouseFactory.deploy()) as TestClearingHouse
+            await testClearingHouse.__TestClearingHouse_init(
                 clearingHouseConfig.address,
                 vault.address,
                 quoteToken.address,
@@ -161,6 +161,7 @@ export function createClearingHouseFixture(
                 accountBalance.address,
                 insuranceFund.address,
             )
+            clearingHouse = testClearingHouse
         } else {
             const clearingHouseFactory = await ethers.getContractFactory("ClearingHouse")
             clearingHouse = (await clearingHouseFactory.deploy()) as ClearingHouse
@@ -176,7 +177,6 @@ export function createClearingHouseFixture(
         }
 
         await clearingHouseConfig.setSettlementTokenBalanceCap(ethers.constants.MaxUint256)
-
         await quoteToken.mintMaximumTo(clearingHouse.address)
         await baseToken.mintMaximumTo(clearingHouse.address)
         await baseToken2.mintMaximumTo(clearingHouse.address)